StandaloneMmCore: Change log level when mCommunicationBuffer is NULL
On Arm all requests are handled as Asynchronous events by the Root
MMI handler.
Since the communication data is conveyed using either the NS shared
buffer or the Secure shared buffer, the Arm implementation does not
setup the mCommunicationBuffer. Therefore, the mCommunicationBuffer
being NULL is not an error case.

Moreover, the existing code switches to Asynchronous event processing
when the mCommunicationBuffer is NULL, which means that the log is an
info log rather than an error.

Therefore, change the log level from ERROR to INFO when the
mCommunicationBuffer is NULL.

Signed-off-by: Levi Yun <yeoreum.yun@arm.com>
(cherry picked from commit 6e3a9027d253753c878f0d6fe94337856e8f82a3)

# Conflicts:
#	StandaloneMmPkg/Core/StandaloneMmCore.c