Add release branch backport workflow (#1201)
## Description

This new workflow will attempt to automatically cherry-pick marked
contributions to a development branch to its corresponding release
branch. If a merge conflict occurs, the commit is committed to a new
branch with merge markers and then a PR is created into the target
branch with those markers. The PR is labeled with
`type:release-merge-conflict` to indicate that it needs manual
resolution.

The PR (if created) is expected to fail compilation and status checks of
course due to the merge conflict markers. A human should then checkout
the PR branch, resolve the conflicts, and push the changes back to the
PR branch.

This workflow is currently being tested in the Mu Basecore repository.
Once it performs well over a reasonable period of time, it will be moved
to the Mu DevOps repository for general use and synced to other relevant
repositories using a more generalized template to fill in some of the
details.# Please enter the commit message for your changes.

---

- To mark a PR going into the development branch so that it should be
cherry-picked to the release branch, add the `type:backport` label to
the PR.
- Note: This label will be added to Mu DevOps to be synced with other
labels when the workflow is added there. For now, it is expected this
label will be manually defined only in Mu Basecore.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

On fork:

- Direct pushes to branch exit successfully.
- Single cherry-pick.
- Multiple cherry-pick.
- PR without label.
- PR with many labels.
- Commits that resolve as empty and need no further action.
- Commits that have merge conflict and need a PR created.

## Integration Instructions

N/A

---

Send feedback to @makubacki during this trial period of the workflow
being used in 202405 Mu Basecore.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>