BaseTools HostBasedUnittestRunner Modify CC_REORGANIZE to exclude Null implementations. (#1218)
## Description

When Unit tests are run, the generated report includes null libraries in
the calculation.

Modifying the organize_coverage to exclude null libraries from the code
coverage results.

These changes only effect the case when CC_REORGANIZE is set to true in
the project.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?
- [x] Backport to release branch?

## How This Was Tested

Ran on private repo which contained a lot of NULL libraries. Prior to
change, Null libraries were included in the report, after the
modification, they were not

## Integration Instructions

No changes required, but coverage report will look different if
CC_REORGANIZE is set to True.